Markdown support and development
Markdown support :
	modified:   controllers/articles.js
	modified:   public/js/articles.js
	modified:   views/articles/create.ejs
	modified:   views/articles/edit.ejs
	modified:   package-lock.json
	modified:   package.json

Changed 'domain' to 'url', removed global cors middleware and added to api routes :
	modified:   config/index.js
	modified:   config/keys.json
	modified:   controllers/auth.js
	modified:   middlewares/auth.js
	modified:   public/js/auth.js
	modified:   routes/api/index.js

Removed redundant repeatation of existing middleware :
	modified:   config/functions.js
	modified:   controllers/articles.js
	modified:   controllers/auth.js
	modified:   middlewares/auth.js
	modified:   routes/articles.js
	modified:   routes/auth.js

Added some utility classes and other styles :
	modified:   public/styles/others/articles.css
	modified:   public/styles/reset.css
	modified:   public/styles/util.css

+++ controllers/articles.js
@@ -1,9 +1,7 @@
-const { verify } = require('jsonwebtoken');
+const marked = require('marked');
 
 const Article = require('../models/Article');
-const User = require('../models/User');
 const { handleErrors, toDate } = require('../config/functions');
-const { code } = require('../config/keys.json').jwt;
 
 const articles_get = (req, res) => Article.find()
   .then((articles) => res.render('articles/home', { title: 'Articles', articles }))
@@ -12,42 +10,44 @@ const articles_get = (req, res) => Article.find()
 const createarticle_get = (req, res) => res.render('articles/create', { title: 'Create Article' });
 const createarticle_post = async (req, res) => {
   const { title, body, slug } = req.body;
-  const token = req.cookies.jwt;
-  verify(token, code, async (err, decodedToken) => {
-    const user = await User.findById(decodedToken.id);
-    try {
-      const article = await Article.create({ 
-        title, body, slug, 
-        author: { id: user.id, name: user.username }, 
-        createdAt: toDate(new Date())
-      });
-      res.status(201).json({ article: article._id });
-    } catch (err) {
-      const errors = handleErrors(err).article;
-      res.status(400).json({ errors });
-    }
-  });
+  const { id, username } = res.locals.user;
+  try {
+    const article = await Article.create({ 
+      title, body, slug, 
+      author: { id, name: username }, 
+      createdAt: toDate(new Date())
+    });
+    res.status(201).json({ article: article._id });
+  } catch (err) {
+    const errors = handleErrors(err).article;
+    res.status(400).json({ errors });
+  }
 }
 
 const article_get = (req, res) => Article.findOne({ slug: req.params.slug })
-  .then((article) => res.render('articles/details', { 
-    title: `${article.title} - ${article.author.name}`, article
+  .then(({ title, body, slug, author, createdAt }) => res.render('articles/details', { 
+    title: `${title} - ${author.name}`,
+    article: { title, body: marked(body), slug, author, createdAt }
   })).catch(() => res.redirect('/articles'));
 
 const editarticle_get = async (req, res) => {
   const { id, username } = res.locals.user;
   const { slug } = req.params;
-  const article = await Article.findOne({ slug, author: { id, name: username } });
+  const article = await Article.findOne({ slug, 'author.id': id, 'author.name': username });
   if (article)
-    res.render('articles/edit', { title: 'Edit Article', article });
+    res.render('articles/edit', {
+      title: 'Edit Article',
+      article: { title: article.title, raw: article.body, body: marked(article.body) }
+    });
   else
     res.redirect('/articles');
 }
 const editarticle_put = async (req, res) => {
   const { title, body } = req.body;
   const { slug } = req.params;
+  const { id, username } = res.locals.user;
   try {
-    const article = await Article.findOne({ slug });
+    const article = await Article.findOne({ slug, 'author.id': id, 'author.name': username });
     article.title = title;
     article.body = body;
     article.save();
@@ -60,13 +60,10 @@ const editarticle_put = async (req, res) => {
 
 const deletearticle = (req, res) => {
   const { slug } = req.params;
-  const token = req.cookies.jwt;
-  verify(token, code, async (err, decodedToken) => {
-    const { id, username } = await User.findById(decodedToken.id);
-    Article.findOneAndDelete({ slug, author: { id, name: username } })
-      .then(() => res.json({ redirect: '/articles' }))
-      .catch(() => res.json({ redirect: '/articles' }));
-  });
+  const { id, username } = res.locals.user;
+  Article.findOneAndDelete({ slug, 'author.id': id, 'author.name': username })
+    .then(() => res.json({ redirect: '/articles' }))
+    .catch(() => res.json({ redirect: '/articles' }));
 }
 
 module.exports = {
